[Paywalls V2] Avoids re-parsing the component tree to handle window insets (#2258)
## Description
This is a continuation of
https://github.com/RevenueCat/purchases-android/pull/2247 & #2252. It
uses the newly introduced `StyleFactoryScope` to properly handle window
insets, in particular in the case of a hero image, in a single pass.

This code is (still) tested by `PaywallComponentDataValidationTests`.

## Motivation
Reducing the number of times we parse the component tree improves
performance. It also allows us to remove a lot of relatively complex
recursive parsing code.