Use fallback API hosts when receiving server down response (#2368)
<!-- Thank you for contributing to Purchases! Before pressing the
"Create Pull Request" button, please provide the following: -->

### Checklist
- [x] If applicable, unit tests
- [ ] If applicable, create follow-up issues for `purchases-ios` and
hybrids

### Description
This PR implements a retry mechanism using fallback URL hosts for
supported requests that fail with a server down error

### Notes
* Only some requests support a fallback API host. Currently only
getCustomerInfo and getProductEntitlementMapping support it.
* `AppConfig` now has a `fallbackBaseURLs` containing the list of hosts,
in order of preference, to use for supported requests.
* When a request's response is a server error (5XX), the same request is
immediately retried with the next host of the array.
* The logic is implemented per-request: that is, if a request fails with
a server-down error for host 1 and then falls back and succeeds for host
2, then all future requests will use again host 1 initially. No state is
held between different requests.
* No fallback retries happen when using a proxy URL

iOS PR counterpart -->
https://github.com/RevenueCat/purchases-ios/pull/4970