[Paywalls v2] Avoid image cropping when combining single-sided margin with Fixed size constraint (#2380)
## Bug
Adding a margin to a single side of an image component with Fixed size
constraints would result in the image being cropped as it would treat it
as padding.

## Fix
Adjusts the size and aspect ratio modifiers of the outer `Box`, to
ensure it grows enough to account for the margin.