Add GitHub pre-commit hooks rule for checking the detekt and metalava (#2378)
### Motivation

Add a GitHub pre-commit hook to optionally run Detekt and Metalava
checks before each commit. While not mandatory, it can be a convenient
way to catch issues early. To use it, simply copy the provided
`pre-commit` file into your `.git/hooks` directory and that's all.

If you face the following error: `The '.git/hooks/pre-commit' hook was
ignored because it's not set as executable`

You can resolve it by running the following command: `chmod ug+x
.git/hooks/*`