Correctly translates variables when a fuzzy match is required. (#2402)
## Bug
In the offerings response, variables could be keyed by just the language
(`de`) while the rest of the strings are keyed by language and region
(`de_DE`).

## Fix
The fix is to apply fuzzy locale matching when looking for the right
variables to use.

I also had to make some other changes to properly set up this scenario
in a test.

Fixes https://github.com/RevenueCat/purchases-kmp/issues/386