[Paywalls] Buttons can now open sheets (#2436)
## Description
Adds full support for buttons opening sheets.
- Adds a `SimpleBottomSheetScaffold` to implement it all. We cannot use
Compose's built-in `BottomSheetScaffold` or `ModalBottomSheet`, as those
are experimental APIs. The `SimpleBottomSheetScaffold` is completely
decoupled from paywalls, so it can be used for other things if needed.
- Sheets are dismissed by `NavigateBack` buttons, the system back
button, and by tapping outside it.
- The paywall content is optionally blurred (if configured on the
dashboard). This is only supported on Android 12 and up, as Compose only
supports this on 12 and up. Below 12, we show a translucent scrim. There
are some less-performant ways we can blur content on older Android
versions, which we can look into (see
[Haze](https://github.com/chrisbanes/haze)), but for now this is fine
imo. #abs


https://github.com/user-attachments/assets/2a1559cf-f0f7-4e93-9e21-12a1b346cb6e