Remove assertion on strings in Customer Center test (#2433)
This assertion was failing every time a new string was added to the
`CustomerCenterConfigData`, meaning an update on the fixture and the
expected object was required.

I get this test is useful if we are getting the JSON from the backend
but being it a fixture, it's just testing that the fixture is correct.