[Paywalls] Margin no longer decreases icon size (#2447)
## Bug
A customer reported an invisible icon.

## Cause
Turned out we incorrectly applied the margin in `IconComponentView`. The
margin would subtract from the size. In this case, the icon was 32dp
wide, and horizontal margin was 16dp, resulting in an invisible
(32-16*2) icon.

## Fix
This PR applies the margin correctly, by wrapping in a `Box` which is
sized according to the icon size + margin.