[Paywalls] Ignores `sheet.background` property in schema (#2451)
## Description
The schema for the new bottom sheet functionality has 2 places to define
a background:
- `sheet.background`
- `sheet.stack.background`
  
After discussion with @joshdholtz we decided to ignore
`sheet.background` and only use `sheet.stack.background`, as that is the
more "natural" place to define this.
  
This also gives us the ability to shape the sheet (rounded corners) for
free. Before this PR, the `sheet.background` would be drawn behind the
sheet "unshaped", causing any shape given to the sheet itself to be
invisible.