Issue when loading promotional offers when products are missing base plans (#2471)
There's an issue when loading promotional offers for target products
that miss a base plan

For example this configuration is valid if the product is old and
configured without base plans in the dashboard:

<img width="728" alt="Screenshot 2025-06-27 at 12 05 36"
src="https://github.com/user-attachments/assets/4d121b68-4214-478c-ad2d-05621ebfc810"
/>

The SDK will fail to load the promotional becasue it looks for the
`product.id`, which in that case would look like `annual_freetrial:p1y`.
On the other hand `product.googleProduct?.productId` has only the
product id `annual_freetrial` so we should be looking for that instead
if we can't find a promotional matching the base plan.

There was also another issue when loading the target product, since we
were picking the first one returned, which is not ideal. For example, if
the product "subscription" had "monthly" and "annual" base plans. And
the target product was just "subscription", `Purchases.getProducts`
without an specific base plan returns two products, one for each base
plan. Which could load an unexpected product. I changed the logic to
load the base plan that's equivalent to the currently active
subscription.