Fix `Switch` component previews (#2509)
### Description
This fixes the previews for the `Switch` component so we can preview the
checked state. They weren't working since they tried to use an unused
property `defaultValue`, which is removed in this PR. The component uses
the selected tab index to know whether the switch should be checked or
not.