Fetch virtual currencies from network (#2518)
This PR includes the following PRs with no other changes:
- https://github.com/RevenueCat/purchases-android/pull/2466
- https://github.com/RevenueCat/purchases-android/pull/2496
- https://github.com/RevenueCat/purchases-android/pull/2505

It is the collection of all of the work required to update the VC APIs
and fetch the virtual currencies from the new v1 endpoint. To keep the
PR sizes manageable, the work was broken up into smaller PRs, which have
been individually approved and aggregated into this branch.

Forgoing a review on this PR since all changes have already been
reviewed, and we'll do a final review when we merge
`virtual-currency-dev` into `main`