Make PaywallActivity not consume insets in Android 16+ (#2520)
### Description
This makes it so we don't consume insets in paywalls in Android 16+.
Starting on Android 16+, by default apps will go edge to edge. We
already consume insets correctly within the paywall as needed, but the
activity was actually consuming those insets before that.

A better fix might be to add an option to enable/disable this
consumption, while keeping the current behavior in Android < 16 but
changing it for android 16+. This is a short term fix for now.