Expose API key validation result to BillingFactory (#2542)
### Description
This is the first of a series of PRs to support a test store in
purchases-android.

In this one, we add support for a new test API key and expose the API
key validation result to the billing factory, so we can in future PRs
create a billing implementation for this test store.