Expose Virtual Currency Constructors with @InternalRevenueCatAPI (#2543)
### Description
This PR exposes the `VirtualCurrency` and `VirtualCurrencies`
constructors publicly behind the `@InternalRevenueCatAPI` annotation.
This will help us in a few scenarios when we need to instantiate VC
objects outside of `purchases-android`, including in the mapper tests in
PHC.

#### Other Approaches Considered
We considered a few other approaches before deciding to use
`@InternalRevenueCatAPI`:
- Making the constructor public without an annotation. This works, but
isn't ideal for a few reasons:
- Adding new properties would be a breaking change unless we used the
`@JvmOverloads` annotation. Even then, we would need to provide a
default value, and that won't always be easy (e.g. when we add VC icons,
these will likely not be optionals)
- Parsing the VC objects from JSON in the PHC mapper tests. This would
work but feels a little dirty, and wouldn't help us if we ever need to
do this in other circumstances.

Using `@InternalRevenueCatAPI` allows us to access the constructor
outside of `purchases-android` while still giving us the freedom to make
breaking changes to the constructor in the future.