Fix insets in Paywall Tester (#2584)
I had noticed wrong padding in the top app bar in the customer center.
After some time trying to understand I came to the conclusion that the
problem was the nested Scaffold. We have one in the MainScreen of
Paywall Tester, and one in the Customer Center.

It was adding the insets to the top of the top app bar, because they
were not being consumed. I thought we had an issue in the Customer
Center but it was actually in the paywall tester. Since the outer
Scaffold was adding a padding, and not consuming the insets, this is how
it was being displayed:

<img width="260" height="575" alt="Screenshot 2025-08-01 at 14 42 32"
src="https://github.com/user-attachments/assets/3642c647-a047-4f01-84cc-fd2af6b3653d"
/>

I removed the padding, and added it to each of the screens as needed,
and consumed it, so the top bar didn't get unconsumed insets and tried
adding them again. This is the final result:

<img width="258" height="575" alt="Screenshot 2025-08-01 at 14 41 49"
src="https://github.com/user-attachments/assets/564db96e-21ae-40ea-8dc8-8ee7bcd01eb0"
/>

I also noticed the bottom bar was not material 3, and the insets were
also not being automatically handled