Fix title and price of non-Google purchases in Customer Center (#2576)
We were displaying the product id for purchases that we don't have the
details for. This happens for example for Apple purchases, or for
Stripe.

<img width="239" height="161" alt="Screenshot 2025-07-31 at 14 12 32"
src="https://github.com/user-attachments/assets/21529c71-7f42-485f-98fa-40a194c121a7"
/>

With this change we will start displaying the entitlement, or if there's
no entitlement, "Subscription" or "One time purchase"