Update to Dokka 2.0.0 (#2609)
### Description
This updates to Dokka 2.0 which fixes some issues with the bump of
gradle version to 9.0.0. I did appreciate some small changes on the
designs, most problematic one might be that the order of rendering seems
to be different... but the docs themselves seem to be mostly the same.
Links seem to work fine and didn't appreciate any issues.