Use `Block store` to backup anonymous user ids across installations (#2595)
### Description
Based on #2594 

This uses [Block
store](https://developer.android.com/identity/block-store) to store the
anonymous user id in the cloud when we find it has any purchases. This
user ID will be recovered and aliased together with any new anonymous
user id when restoring or completing a purchase.

This will be useful to avoid issues when trying to restore purchases
with consumed consumables that were meant to never be consumed.