Fixes price formatting discrepancies on Paywalls for `{{ product.price_per_[day|week|month|year] }}` (#2604)
## Bug
We've had reports of formatting discrepancies between calculated prices
(e.g. monthly price calculated from a yearly subscription) and the price
that comes directly from the store (the yearly price of a yearly
subscription). This store price is shown on the paywall (`{{
product.price }}`) and on the Play Billing purchase sheet.

## Cause
This is caused by the fact that calculated prices use the device locale
to format, and the store uses the storefront country.

## Fix
The fix implemented in this PR is to use the storefront country to
format all prices. The difficulty is in the fact that the storefront
country is only a country, not a complete locale. It's missing a
language. To solve that, we check all available device locales for the
storefront country, and pick the best one.