Revert dokka 2 and gradle 9 update (#2618)
### Description
We've been having some issues with Dokka 2. Our dokka-hide-internal
plugin tests are not passing with not a great description. We've decided
to revert the update for now. Unfortunately, that also means we need to
revert the Gradle 9 update, to keep it compatible with Dokka 1.

This PR also changes to run the dokka-hide-internal plugin tests on all
branches so we don't miss this again.