Migrate RC SharedPreferences data to a separate file (#2621)
### Description
We were storing data in the default shared preferences file, which could
potentially be used by the app developer. That's a bad practice and it's
more difficult to do some things with. This PR changes to a separate
shared preferences file and adds a migration, so we migrate all RC data
from the previous shared preferences file to the new one. We won't
delete the data from the previous shared preferences file, but that will
remain unused.