Handle payment pending errors when restoring properly (#2635)
### Description
Until now, if trying to restore a purchase with a payment pending error,
the SDK returned an `UnknownBackendError` code. This should be returned
instead with a `PaymentPendingError` code for proper handling.

Note that restore errors return the error of the last post receipt with
errors. If there were multiple errors, all might not be exposed right
now.