MON-1193 Support delayed close button (Component Transitions) (#2623)
<!-- Thank you for contributing to Purchases! Before pressing the
"Create Pull Request" button, please provide the following: -->

### Checklist
- [x] If applicable, unit tests

### Motivation
We have several users who want the ability to delay the appearance of
the close paywall button.


### Description

Implemented a Transitions mechanism that will optionally hide/show a
view when the component is marked with a transition object. This lays
the groundwork for accepting custom animation declarations so that it
would not error out if/when we add that in the future.