Video Component Models (dark code) (#2646)
<!-- Thank you for contributing to Purchases! Before pressing the
"Create Pull Request" button, please provide the following: -->

### Checklist
- [x] If applicable, unit tests
- [ ] If applicable, create follow-up issues for `purchases-ios` and
hybrids

### Motivation

Creating a new paywall component can be a VERY large amount of code, I'm
breaking things apart to make the PR's easier to review. This PR is just
the creation of the VideoComponent model.

### Description

Add a video component model and serialization tests


> [!Note]
> I did not make it a member of the PaywallComponent sealed interface. I
intentionally left that part out so I wouldn't have to have a style
factory function setup and all of the other stuff that would domino into
the PR.
> This is just the json -> model