File Repository (#2634)
<!-- Thank you for contributing to Purchases! Before pressing the
"Create Pull Request" button, please provide the following: -->

### Checklist
- [x] If applicable, unit tests
- [ ] If applicable, create follow-up issues for `purchases-ios` and
hybrids

### Motivation

We will be supporting Video components in the Paywalls soon, to do that
we need a good way to cache the large files to the device.

### Description

This PR is a carbon copy of what I did in the iOS sdk recently. There
are 2 cache mechanisms.

1. A deferred value cache—This essentially allows us to reuse the same
coroutine job across multiple call-sites. That way we can prefetch and
move on. When the paywall is launched the same download can be awaited
when it's partially completed. Until it finishes, we can show the
fallback image that should download far faster.

2. The file repository—This knows when to fetch the data from the web or
when to read it from the file system. There are likely things I've
missed here. But there are a couple of important things to note.
a. We are not currently hosting the videos in our ecosystem so a
checksum isn't realistic in our v1
b. I'm good at Kotlin, but not at all experienced with Android SDK
development . Totally open to feedback.

(cherry picked from commit 6501830b094981f0b432345429c722e9050272e9)