Fix bundled font families when using assets folder (#2672)
### Description
As a follow-up from
https://github.com/RevenueCat/purchases-android/pull/2624.

That PR fixed embedded font families when they were in the resources
folder, as it's common in android. However, that didn't fix the issue
when fonts are in the assets folder. For example, when being added in
React native when using `npx react-native-asset`.

This will fix that so font families load correctly when using the assets
folder.

(cherry picked from commit 3b806eda790bb38bb9bb4c4ab7a914a37e007743)