Finalize video component to include a checksum, and optimize memory usage for large file downloads (#2717)
<!-- Thank you for contributing to Purchases! Before pressing the
"Create Pull Request" button, please provide the following: -->

### Checklist
- [x] If applicable, unit tests
- [ ] If applicable, create follow-up issues for `purchases-ios` and
hybrids

### Motivation
<!-- Why is this change required? What problem does it solve? -->
<!-- Please link to issues following this format: Resolves #999999 -->

After getting the last video component PR up, the team had some great
feedback, specifically that there was no way to prevent the corruption
of a file before saving it and that storing the whole file in memory
could be problematic when dealing with larger files like videos. So this
PR sought to solve those issues.

Description

This PR does several things:

- Stores the file using the checksum value if present -- ensuring that
we get a cache miss if the checksum value changes
- Streams the download into a temporary file to optimize memory usage
for very large downloads (as opposed to keeping the whole thing in
memory)
- Only saves the data to the cache directory if the checksum matches the
dowloaded content
- Stops the cache warming of the videos to save on costs - we will be
adding a flag in our backend to denote which paywalls can prewarm videos
in the future
- Removes the static object responsible for state management and
replaces it with a view model to survive rotation

### Description
<!-- Describe your changes in detail -->
<!-- Please describe in detail how you tested your changes -->
# Conflicts:
#	purchases/src/main/kotlin/com/revenuecat/purchases/paywalls/components/properties/VideoUrls.kt