Protect against developers excluding blockstore dependency (#2745)
### Description
We got a report of some crashes trying to find the Blockstore
dependency. Looks like it was excluded from the runtime dependencies.
While this seems like an invalid usage, the Blockstore dependency is not
mandatory for usage of the SDK, so we can protect against it being
excluded, same way we protect against the com.google.android.gms ads
identifier library being excluded.

I've tested this by excluding the dependency and it doesn't crash with
these changes.