Add webCheckoutURL property to Offering and Package (#2747)
### Description
This adds a new API for both `Offering` and `Package`, the
`webCheckoutURL`. This property will have the Web Purchase Link
associated with that Offering and/or Package, so it can be opened if
needed. This will be needed to support Web Purchase Buttons in Paywalls,
coming in a follow-up PR.