[Experimental]: Support Multi-Line Subscriptions (#2668)
### Description
Adds support for purchasing add-on subscriptions in the
`purchases-android` library. Through the APIs introduced in this PR,
developers can purchase subscriptions with add-ons via a new
`PurchaseParams.Builder` function that accepts a `List<StoreProduct>` or
`List<Package>` of the add on products/packages that the user would like
to purchase:

```kotlin
@OptIn(ExperimentalPreviewRevenueCatPurchasesAPI::class)
fun purchaseSubscriptionWithAddOns() {
    val purchaseParams = Builder(
        activity = activity,
        storeProduct = baseProduct
    )
        .addOnProducts(listOf(product2, product3))
        .build()

   val purchaseParams = Builder(
        activity = activity,
        packageToPurchase = basePackage
    )
        .addOnPackages(listOf(package2, package3))
        .build()

    Purchases.sharedInstance.purchase(
        purchaseParams = purchaseParams,
        callback = // Omitted for brevity
    )
}
```

Additionally, this PR introduces a new property to `StoreTransaction`:
`val subscriptionOptionIdForProductIDs: Map<String, String>?`. This
property contains a mapping of productIDs to the subscription option ID
that was purchased. This is helpful in two scenarios:
- It allows developers to know which subscription options were purchased
in a multi-line subscription
- It allows the SDK to include subscription option IDs for multi-line
add-ons when POSTing the receipt to the RC backend.


### Notes
The new `PurchaseParams` function and
`StoreTransaction.subscriptionOptionIdForProductIDs` are marked as
`ExperimentalPreviewRevenueCatPurchasesAPI`, so it will require
developers to opt-in to using an experimental API, as shown in the code
snippet. We can remove in the future when the feature is more stable.