Add more proper error message when using test store API key in release builds (#2762)
### Description
We were using the ConfigurationError code message and passing the
details in the underlying error message. However, this made it so the
proper error isn't that accessible, specially for hybrids developers.
This modifies the exception message to be more precise instead of
leaving the generic one.

This is a different approach than the one in #2752