Remove GOOGLE_PRODUCT_MISSING_INFINITELY_RECURRING_BILLING_PHASE log message (#2767)
### Description
Removes the `GOOGLE_PRODUCT_MISSING_INFINITELY_RECURRING_BILLING_PHASE`
log message. This was intended to log a warning message whenever we
couldn't find an infinitely recurring billing phase on a Google
subscription product. We thought this wouldn't happen often, but it
happens for all prepaid subscriptions, since they don't renew
automatically. This was creating a bit of noise in the logs when working
with prepaid subscriptions.

This isn't a problem for our processing of subscriptions with add-ons,
since subscriptions with add-ons don't support prepaid subscriptions.