Change test server down mechanism to use new server down endpoint (#2765)
### Description
In order to test the server being down, we had a mocked response in our
integration test. However, this didn't test the exact behavior of the
backend. This PR just changes the mechanism we use to test these
situations to instead use a specific endpoint that will always behave as
if the server was down.