[EXTERNAL] Update Emerge snapshots to best/latest practices (#2777) by @rbro112 (#2779)
### Checklist
- [ ] If applicable, unit tests
- [ ] If applicable, create follow-up issues for `purchases-ios` and
hybrids

### Motivation
While working on solving a bug yesterday, I noticed Emerge snapshots are
a few versions behind/not using latest best practices. This updates
that!

### Description
Updates a few things to now follow best practices:
- Updates Emerge Gradle plugin to latest (4.4.0)
- Updates Emerge snapshots to latest (1.5.0)
- Replaces `emerge-snapshots-annotations` (deprecated) with
`emerge-snapshots-runtime`.
- With above library change, updates to use latest
`@EmergeSnapshotConfig` annotations in place of `IgnoreEmergeSnapshot`
- Sets `baseSha` to `null` on non-PR builds to avoid main branch
failures, as `""` to us is a valid baseSha, which causes us to search
for a non-existing build to compare against. This should cause main
branch runs to show a green check now.

Contributed by @rbro112 in #2777

---------

Co-authored-by: Ryan Brooks <ryanjeffrey.brooks112@gmail.com>