Unlock entitlements from add-ons in offline entitlements mode (#2776)
### Description
Adds support for unlocking entitlements from products purchased as a
subscription add-on. This is done by modifying the
`OfflineCustomerInfoCalculator` to consider all product IDs in a
`StoreTransaction` when calculating the purchased products, instead of
just the first one.

The PR has the `pr:other` tag since support for subscriptions with
add-ons is still experimental.