Add extra request logging (#2794)
### Description
From time to time, I've found myself wanting to get a curl request for
the requests performed by the SDK during a test run, which makes it
annoying to get it through the network inspector, since you need to run
while debugging to be able to get them to show and it's a bit annoying.
In those situations, I've added this small piece of code to log the curl
request + full response. I thought it could be useful to have it in the
repo but wanted to gather thoughts first. Wdyt?