Unify integration tests jobs (#2791)
### Description
This PR unifies the existing integration test jobs for production and
load shedder into a single one, with some internal code in fastlane to
decide what environment variables to use. In future PRs, we will change
the environment variables to try to unify as much as possible, and only
split where needed.