Video Background (#2786)
<!-- Thank you for contributing to Purchases! Before pressing the
"Create Pull Request" button, please provide the following: -->

### Checklist
- [ ] If applicable, unit tests
- [ ] If applicable, create follow-up issues for `purchases-ios` and
hybrids

### Motivation
<!-- Why is this change required? What problem does it solve? -->
<!-- Please link to issues following this format: Resolves #999999 -->

Make it easy to add videos to a paywall.

### Description
<!-- Describe your changes in detail -->
<!-- Please describe in detail how you tested your changes -->

Rendering the video as a background requires that it be in a z-stack and
THAT seems to require very explicit sizing for android to actually
render the video component at all. The only way I could figure out how
to do that and support the custom layout that we're doing for buttons
and no-badge stacks was to essentially duplicate the sizing logic that
happens in side of the stack function. If we apply those same modifiers
on the optional video view when it's not a video, it's not put in a box,
and then we get a crash for making an infinitely scrolling view. So I
have wrapped it in another check for now. It's ugly, and I know it, so
I'd like someone more skilled with Compose to look at it.



https://github.com/user-attachments/assets/d4537f17-63ee-4806-a3fa-77e5ba3856c7