Add hook to delay payment screens (#2753)
<!-- Thank you for contributing to Purchases! Before pressing the
"Create Pull Request" button, please provide the following: -->

### Checklist
- [ ] If applicable, unit tests
- [ ] If applicable, create follow-up issues for `purchases-ios` and
hybrids

### Motivation

Some of our customers have requested a way to intercept the on purchase
flow before the purchase happens so they can ensure other critical steps
are done first like authentication, or age verification.

### Description
<!-- Describe your changes in detail -->
<!-- Please describe in detail how you tested your changes -->

Creates the ability to gate and intercept when a purchase / redeem
button is pressed. This should allow the consumer to log or intercept
the flow and present an auth flow for example.