Add `bc8` default flavor (#2829)
### Description
First step to split bc7 and bc8 flavors. For now, just creating an empty
default bc8 flavor, which will be used by all code. In follow-up PRs, I
will start moving code to the correct flavor, before applying the
changes in the pretty big #2816 . This PR is pointing to a base branch
to merge all PRs, not to main.

No changes should happen as consequence of this PR.