Countdown Paywalls component (#2837)
Equivalent of
([#5790](https://github.com/RevenueCat/purchases-ios/pull/5790)).

### Overview
Adds support for countdown timers in paywalls that can display different
content based on whether the countdown is active or has ended.

### Note on prewarming vides/images
When a countdown ends and transitions from `countdownStack` to
`endStack`, any images or videos in the end stack (which can contain
anything) need to be immediately available.


[Screen_recording_20251113_125835.webm](https://github.com/user-attachments/assets/9f794ab6-b848-45af-a847-d77cd10d7dad)