CC-581 | Allow for support ticket creation (#2810)
### Motivation

This will allow customers to create a support ticket in order to contact
the developer, rather than directly emailing the developer.

### Description

- If the correct config is passed through (from enabling support tickets
in the dashboard) the "Contact us" button will no longer open the native
mail client, but will instead show a new screen. This screen allows
customer to give their email and a description of the issue. This will
then be passed on to the developer from the backend.

I tested my changes locally (emulator only, due to no Android device)
both against my local API and the production API, using existing CC
configs and some with the new changes. I also added unit tests for the
new screen.

[iOS PR](https://github.com/RevenueCat/purchases-ios/pull/5779)