Add client side timeout logic for endpoints that support fallback URLs (#2807)
### Description
This adds some logic to implement a client side connection timeout. The
logic is like this:
- If the request supports fallback URLs and we're attempting the "main"
backend, timeout will be 5s.
- If the request is performed to a fallback URL or does not support
fallbacks, timeout will be 30s
- Once we timeout in one of the endpoints that support fallbacks, the
new timeout for these endpoints that support timeouts in the main server
will be 2s, so we can fail early and try the fallback.
- All requests will always hit the "main" backend first.
- Whenever the main server is hit successfully, the timeout is reset to
the initial 5s for the endpoints that support fallbacks

This is pretty delicate, so lmk what you think!

---------

Co-authored-by: Antonio Pallares <ajpallares@users.noreply.github.com>