[MON-1122] Revert variable rounding logic to not round up (#2857)
When shipping Paywalls v2, there was a regression with the way that we
show prices. We changed from not rounding or rounding down to actually
rounding up.

Where this is specially important is when a dev picks a price that is
explicitly designed to get just under a dollar threshold, like $83.99,
which is either 6.99 or 7.00 — and 6.99 is the actual price they're
aiming to put in front of customers.

### Checklist
- [x] If applicable, unit tests
- [x] If applicable, create follow-up issues for `purchases-ios` and
hybrids