Runs plugin actions from correct directory (#2858)
## Description
We ran into the Fastlane quirk where plugin actions run from the
**parent** directory of the lane's current working directory. This is
only a problem when the lane's current working directory is not the
`/fastlane` directory, but e.g. the repository root. Because this is not
always a problem, I did not want to modify the plugin action to accept a
directory parameter. Instead, I added a workaround to the call site.

https://docs.fastlane.tools/advanced/fastlane/#directory-behavior