Fix ANR when initializing FontLoader (#2819)
Reported here
https://github.com/RevenueCat/purchases-android/issues/2818

I believe the cause is we are doing `context.cacheDir` on initialization
(in main thread). Looking at the stacktrace it looks like it does some
checks on the disk on main thread that could be causing ANRs.

I tested with StrictMode and after this fix, the warning is gone.