Extract parameters for non paid revenue tracking API to use objects (#2871)
### Description
This is a follow-up from #2728 

In iOS, we're passing an object to the track methods for non paid
revenue tracking, but we were passing the individual parameters in
Android. After thinking about it, passing an object allows for better
extensibility, so modifying the API for android.

This is technically a breaking change, but the API is still internal, so
it should be ok.