Change paywall preview recorder name template (#2923)
### Description
Looks like in Paparazzi 2.0.0-alpha, the current diamond bracket symbol
is not allowed. Instead, using triple underscores.

This is extracted from #2446, but wanted to do this in a separate PR so
we make sure the templates don't change in that PR when rendering them,
which was harder to see with the rename in Emerge.