Add AIDL to query purchase history inapp products (#2872)
### Description
With BC8, we lost the ability of querying consumed one time products,
which sometimes we needed to unlock entitlements correctly on a restore,
in cases where the products were accidentally consumed. This uses the
old AIDL service to be able to query those products instead of using the
billing library.