Make close() method also clear the shared instance of the SDK (#2940)
### Description
When calling the `close` function, we cleared a bunch of internal state
that leaves the SDK in an inconsistent state. Specifically, we clear the
billing client from the BillingWrapper, among other things. If after
that, you tried to reconfigure the SDK with the exact same
configuration, it wouldn't let you, since we have logic to prevent
reconfigurations with the same config. This means that in some cases, we
would end up in this inconsistent state, without a great way to get out
of it.

I think the `close` function means the instance shouldn't be called
more, so we should also clear the shared instance, so we can actually
reinitialize later if needed and `Purchases.isConfigured` is false. This
is however a behavior change.