Add size testing app and related CI tasks (#2956)
### Description
This adds a new app, straight out of the AS template that will be used
to properly verify the size of the SDK in release builds. Right now the
app we use for these tests is built in debug mode which won't properly
reflect the final size of the SDK and is also used for snapshots that
would be stripped aways from the app (once we enable proper
minification).

---------

Co-authored-by: JayShortway <29483617+JayShortway@users.noreply.github.com>